feat(packaging): cleaner module & image of griffon

+++ scripts/preremove.sh
@@ -1,5 +1,8 @@
 #!/bin/sh
 set -e
 
-systemctl stop griffonav-daemon || true
-systemctl disable griffonav-daemon || true
+if command -v systemctl >/dev/null 2>&1; then
+    systemctl stop griffonav-daemon || true
+    systemctl disable griffonav-daemon || true
+    systemctl daemon-reload || true
+fi
